strava/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
stravalib==1.7